refactor for debugging
