+++ lib/plugin.dart
@@ -32,18 +32,18 @@ import 'flutter/create_project.dart';
 import 'flutter/flutter_launch.dart';
 import 'flutter/toolbar.dart';
 import 'impl/changelog.dart';
-import 'impl/cli_launch.dart';
-import 'impl/console.dart';
 import 'impl/editing.dart' as editing;
 import 'impl/errors.dart';
 import 'impl/outline.dart';
 import 'impl/pub.dart';
 import 'impl/rebuild.dart';
-import 'impl/run_app.dart';
-import 'impl/shell_launch.dart';
 import 'impl/smoketest.dart';
 import 'impl/status_display.dart';
 import 'jobs.dart';
+import 'launch/cli_launch.dart';
+import 'launch/console.dart';
+import 'launch/run_app.dart';
+import 'launch/shell_launch.dart';
 import 'launch.dart';
 import 'linter.dart' show DartLinterConsumer;
 import 'projects.dart';
